Back to normal schedule fragment. Filters the list based on day and displays them in the 3 tabs. 3 network calls as well, yay. Oh well.
