+++ mobile/Client/Client.Droid/Adapters/SchedulePagerAdapter.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
 using Android.Support.V4.App;
 using Client.Droid.Screens;
 using Common.Common.Models;
@@ -21,25 +19,12 @@ namespace Client.Droid.Adapters
 
 		public override Fragment GetItem(int position)
 		{
-			ScheduleFragment fragment = new ScheduleFragment();
-			var list = GetDayEvents(position);
-			fragment.DayEvents = list;
-			return fragment;
-		}
-
-		public override int GetItemPosition(Java.Lang.Object @object)
-		{
-            return PositionNone;
+			return new ScheduleFragment(position);
 		}
 
 		public override ICharSequence GetPageTitleFormatted(int position)
 		{
 			return new Java.Lang.String(titles[position]);
-		}
-
-		private List<Event> GetDayEvents(int day) {
-			var eventsByDay = Events.GroupBy(x => x.StartTime.Day).ToArray();
-			return eventsByDay[day].ToList();
 		}
 	}
 }
