+++ mobile/Client/Client.Droid/Screens/ScheduleFragment.cs
@@ -1,48 +1,76 @@
 ﻿using Android.Support.V4.App;
 using Android.Content;
 using Android.OS;
-using Android.Runtime;
-using Android.Util;
 using Android.Views;
 using Android.Widget;
 using Android.Support.V7.Widget;
 using Client.Droid.Adapters;
-using Common.Common;
 using Client.Common;
 using Common.Common.Models;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Timers;
+using System.Linq;
 
 namespace Client.Droid.Screens
 {
 	public class ScheduleFragment : Fragment, IScheduleView
 	{
 		SchedulePresenter Presenter { get; set; }
-		public List<Event> DayEvents { get; set; }
+
+
 		List<Event> IScheduleView.Events
 		{
 			set
 			{
-				return;
+				ScheduleAdapter adapter = this.Adapter;
+				if (adapter == null) return;
+				adapter.Events = GetDayEvents(value, this.day);
+				adapter.NotifyDataSetChanged();
 			}
 		}
 		RecyclerView ScheduleView { get; set; }
 		ScheduleAdapter Adapter { get; set; }
+		int day;
+
+		Timer timer = new Timer(TimeSpan.FromSeconds(10).TotalMilliseconds) { AutoReset = true };
+
+		public ScheduleFragment(int dayIndex)
+		{
+			day = dayIndex;
+		}
 
 		public override void OnCreate(Bundle savedInstanceState)
 		{
 			base.OnCreate(savedInstanceState);
 
-			Presenter = new SchedulePresenter(null);
+			Presenter = new SchedulePresenter(((ShowdownClientApplication)this.Activity.Application).BackendClient);
 			Presenter.TakeView(this);
 
 			Adapter = new ScheduleAdapter()
 			{
-				Events = DayEvents
+				Events = new List<Event>()
 			};
 			Adapter.ItemClick += (object sender, ScheduleAdapterClickEventArgs args) => Presenter.OnClickRow(args.Event);
+		}
+
+		public async override void OnResume()
+		{
+			base.OnResume();
+
+			Presenter.TakeView(this);
+			await Presenter.OnBegin();
+
+			timer.Elapsed += (sender, e) => Activity.RunOnUiThread(async () => await Presenter.OnTick());
+			timer.Start();
+		}
+
+		public override void OnStop()
+		{
+			base.OnStop();
+			timer.Stop();
+			Presenter.RemoveView();
 		}
 
 		public override View OnCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
@@ -72,6 +100,12 @@ namespace Client.Droid.Screens
 		Task IScheduleView.ScheduleReminder(Event eventToRemind)
 		{
 			return Task.CompletedTask;
+		}
+
+		private List<Event> GetDayEvents(List<Event> e, int day)
+		{
+			var eventsByDay = e.GroupBy(x => x.StartTime.Day).ToArray();
+			return eventsByDay[day].ToList();
 		}
 	}
 }
